Setup CircleCI for master branch and Pull Requests
To setup circleci i created a Makefile, i moved to code to a subfolder in order to clean up the codebase, so we know non-pony setup code and config files etc. lives at the root and the pony code live in reactive-streams subdir.

The circleci setup and the restructuring was heavily influenced by https://github.com/ponylang/pony-stable